A*, Nobbins, Threading, and Frame-rate. Oh My!
Okay, so Nobbin now implements A*(Overkill) to find shortest path to
hero, needs a bit of final touches for path reconstruction and such.
We now have a Update runnable that we will make into a thread that
updates all entities in sync (currently set to 30 FPS). Entities will
have to be changed slightly to incorporate that, basically instead of
things calling move, they will set dx/dy to some value, and when they
are told to update it will check that enough time has passed that it can
act again, if so move it those dx and dy values and set dx and dy back
to 0. Framework is there, going to try to finish up things to have
something we can 'show off' in the morning, committing now incase I fall
asleep.