GOLD! Fellowship Get! Broken = Rotate and Flatten?
SO MANY FILES.
Made some art for the gold(fellowship), we can just add a random number
to the end of the sprite name to get the randomized fellowship or we can
have each level have 8 for the whole group.
Slowed down gold decent and worked on switching for gold. Got rid of
Enemies class because the way we did things we didn't need switch.