Complete Game?
Added a Start menu that allows selection between the game and level
editor.

Added Gold finding to Hobbin.

Added enemySpawn point to levels. (for level editor it is the last
placed Nobbin)

Added a few java docs.

There are 3 lines in level that can be commented out to stop the
randomized testing game and go straight to pre-made levels.